+++ packages/generator-teams/src/messageExtension/templates/src/server/{messageExtensionName}/__tests__/{messageExtensionClassName}.spec.ts
@@ -1,0 +1,8 @@
+import <%=messageExtensionClassName%> from "../<%=messageExtensionClassName%>";
+
+describe("<%=messageExtensionClassName%>", () => {
+    it("Should create the middleware", () => {
+        const middleware = new <%=messageExtensionClassName%>();
+        expect(middleware).toBeDefined();
+    });
+});