+++ packages/generator-teams/tests/tabGenerator.spec.ts
@@ -197,10 +197,10 @@ describe("teams:tab", function () {
 
     if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
       const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
-      assert.equal(false, npmInstallResult);
+      assert.strictEqual(true, npmInstallResult);
 
       const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
-      assert.equal(false, npmRunBuildResult);
+      assert.strictEqual(true, npmRunBuildResult);
     }
 
     await helpers
@@ -222,10 +222,10 @@ describe("teams:tab", function () {
 
     if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
       const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
-      assert.equal(false, npmInstallResult);
+      assert.strictEqual(true, npmInstallResult);
 
       const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
-      assert.equal(false, npmRunBuildResult);
+      assert.strictEqual(true, npmRunBuildResult);
     }
   });
 
@@ -288,10 +288,10 @@ describe("teams:tab", function () {
 
     if (process.env.TEST_TYPE == testHelper.TestTypes.INTEGRATION) {
       const npmInstallResult = await testHelper.runNpmCommand("npm install --prefer-offline", projectPath);
-      assert.equal(false, npmInstallResult);
+      assert.strictEqual(true, npmInstallResult);
 
       const npmRunBuildResult = await testHelper.runNpmCommand("npm run build", projectPath);
-      assert.equal(false, npmRunBuildResult);
+      assert.strictEqual(true, npmRunBuildResult);
     }
 
   });